# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matchbook API",
    "description": "Self-hosted agency finance ops \u2014 invoice collection, reconciliation, reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/reconciliation/upload": {
      "post": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Upload Statement",
        "description": "Upload a bank statement file, parse via LLM, store transactions.",
        "operationId": "upload_statement_reconciliation_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_statement_reconciliation_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/overview": {
      "get": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Period Overview",
        "description": "Full reconciliation overview for a period \u2014 shows whether the month is complete.",
        "operationId": "period_overview_reconciliation_overview_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Period to reconcile (YYYY-MM)",
              "title": "Period"
            },
            "description": "Period to reconcile (YYYY-MM)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodReconciliation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/transactions": {
      "delete": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Delete All Transactions",
        "description": "Delete all transactions (and their matches), reset linked invoice statuses.",
        "operationId": "delete_all_transactions_reconciliation_transactions_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete All Transactions Reconciliation Transactions Delete"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_reconciliation_transactions_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransactionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/transactions/{transaction_id}/dismiss": {
      "post": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Dismiss Transaction",
        "description": "Dismiss a transaction \u2014 mark as no_invoice with an optional note (e.g. 'currency conversion fee').",
        "operationId": "dismiss_transaction_reconciliation_transactions__transaction_id__dismiss_post",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDismiss"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/match": {
      "post": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Trigger Matching",
        "description": "Trigger hybrid matching: deterministic first, then LLM for leftovers.",
        "operationId": "trigger_matching_reconciliation_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/matches": {
      "post": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Create Match",
        "description": "Manually create a match between an unmatched invoice and transaction.",
        "operationId": "create_match_reconciliation_matches_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "List Matches",
        "operationId": "list_matches_reconciliation_matches_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/matches/{match_id}": {
      "get": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Get Match",
        "operationId": "get_match_reconciliation_matches__match_id__get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Reject Match",
        "description": "Reject (delete) a match \u2014 resets invoice and transaction to unmatched.",
        "operationId": "reject_match_reconciliation_matches__match_id__delete",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reconciliation/matches/{match_id}/confirm": {
      "post": {
        "tags": [
          "reconciliation",
          "reconciliation"
        ],
        "summary": "Confirm Match",
        "description": "Confirm a match \u2014 updates invoice and transaction status.",
        "operationId": "confirm_match_reconciliation_matches__match_id__confirm_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_statement_reconciliation_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_statement_reconciliation_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchTriggerRequest": {
        "properties": {
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "type": "object",
        "title": "MatchTriggerRequest"
      },
...
      "StatementUploadResponse": {
        "properties": {
          "transactions_parsed": {
            "type": "integer",
            "title": "Transactions Parsed"
          },
          "period": {
            "type": "string",
            "title": "Period"
          }
        },
        "type": "object",
        "required": [
          "transactions_parsed",
          "period"
        ],
        "title": "StatementUploadResponse"
      },
...
      "TransactionDismiss": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "TransactionDismiss"
      },
      "TransactionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TransactionList"
      },
...
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "unmatched",
          "matched",
          "no_invoice",
          "withholding"
        ],
        "title": "TransactionStatus"
      },
...
    }
  }
}